- This PR make the code golint and gofmt -s compliant.
- Removes all the `fmt` calls and replace with faster alternative.
- Use setters for global namespace and tags
- Use io.WriteCloser instead of net.Conn to allow for easy Mock / Extension

Benchmark before:
```
BenchmarkGauge           1000000              1302 ns/op
BenchmarkCount           2000000               834 ns/op
BenchmarkHistogram       1000000              1283 ns/op
BenchmarkSet             2000000               823 ns/op
BenchmarkEvent           1000000              1318 ns/op
BenchmarkSend            3000000               516 ns/op
```

After:
```
BenchmarkGauge           2000000               727 ns/op
BenchmarkCount          10000000               214 ns/op
BenchmarkHistogram       2000000               728 ns/op
BenchmarkSet            10000000               143 ns/op
BenchmarkEvent           3000000               478 ns/op
BenchmarkSend           20000000               108 ns/op
```