@irabinovitch 2015-10-12
@mananchandra Thoughts on merging this?